Break out action-api
