+++ packages/action-api/src/ActionExecutor.ts
@@ -1,0 +1,6 @@
+import { Result } from '@chellimiller/result-api'
+
+/**
+ * Function to execute an action.
+ */
+export type ActionExecutor<DATA, ERROR> = () => Promise<Result<DATA, ERROR>>
