+++ packages/action-api/src/ActionSeverityType.ts
@@ -1,0 +1,9 @@
+/**
+ * Describes the severity of an action. Allows differentiation between most actions and potentially dangerous actions.
+ */
+export enum ActionSeverityType {
+  /** Applies to most actions. */
+  STANDARD = 'STANDARD',
+  /** Taking this action could potentially have negative consequences. For example, deleting an object. */
+  DANGER = 'DANGER',
+}
